Update pattern handle cell on construction and click only
The chosen cell of the pattern for editing should only be selected when
constructing the pattern handle holder, or when the user clicks on a
space within the pattern. All other times it should remain constant.

This means the cell will not change during regular updates, for example
when dragging the size around.

Also fixes an issue with clicking on grouped patterns

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3868
Fixes https://gitlab.com/inkscape/inkscape/-/issues/4313