Fix several includes to 2geom headers
For whatever reason, there are several 2geom `#includes` that are in
the form `#include "foo.h"` instead of `#include <2geom/foo.h>`.

On my system, this was causing compile errors, but it doesn't seem
to be causing issues for the CI, I'm not sure why.

Regardless, includes like that are fragile, and it's probably best to
replace them with proper ones.