Fix bad mask and clip editing
When the mask or clipping objects have their own transformations the
node editor incorrectly positions the editing node handles and it's
hard to use or see correctly.

This is because the clipObject's i2d_affine must be applied ON TOP of
the base item's i2d_affine, not under it.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/711