Create new Extension Processing Actions features
We create a new system of document level actions which modify the
document during safe (and in the future, effect extensions too) these
can do various processing steps such as object to path, or text to path
and should provide us with a large amount of future flexibility.

The processing steps which were embedded within the svg saving are moved
to document actions and the actions are added to the internal svg output
extensions.

To keep compatibility we also add a feature to enable and disable the
processing steps through the system wide preferences.