Use GitLab's fleet to run Windows CI
The build process is divided into build and dist jobs in order to speed
up the overall process. It's also a way to mitigate current limitations
like a runner-imposed limit of 120 minutes runtime.

The test job hasn't been active for a while and is not implemented yet.
It'll make a return later.

Only 7z and exe distributions are being built with msi to be implemented
later.

The build job depends heavily on ccache to make up for the bad
performance of GitLab's Windows VMs. In fact, we wouldn't be able to
complete a single build of Inkscape without it. The other invaluable
component is the restart mechanism (`retry: 2` setting) that is being
used to split up the load from a full/fresh build (read: with few or
zero cache hits) into multiple runs.