PDF Files sometimes contain UTF16 text and must be converted
We use g_convert to turn detected utf16 (BE and LE) into UTF8,
this is based on code from poppler/glib which must do a similar
conversion to GValues but we convert to std::string.

Fixes PDF files with broken and corrupt label names which cause
many problems and sometimes crashes.