Fix bad export of overlapping clipped layers
Where multiple layers existed, and some of them have clipped contents which would
overlap other pages. Parts of those layers were being exported in the pdf/ps
file types because the layer with the clip was being stripped out.

We change our overlapping function to allow us to control layer selection and thus
include the layer in the selection when rendering.