Adds raster fracturing and canvas opacity controls
Add the ability to generate clipped images from the shape builder fracture process.
Allows the user to set an opacity for the canvas that sits behind the selected objects.

Objects selected are still invisible except for images which are retained.

This accounts for clones with a clipPath attached so they can be used in secondary shape builder actions.

A feature to stop deleting the original items on request, some users want to run shape builder multiple times.