Limit Extension::prefs() calls to GUI execution
The Inkscape::Extension::Extension::prefs() call will spawn a GUI dialog
asking for preferences. The dialog waiting for user inputs interrupts
the execution of CLI.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3524
by checking whether GUI is used in prefs() and skip if it is not.

Removed documentation of a param that seems to be removed.