io/file: can't forward-declare RefPtr now that it…
…in GTK4 is an alias for std::shared_ptr, which (a) could have been
included indirectly already by any number of std:: header files, and (b)
we are not allowed to forward-declare things in the stdlib anyway.

This is not needed in GTK3 but should not hurt much; refptr.h is
unlikely to be much heavier than any header already included here, and
it's one less thing we'll have to think about later when moving to GTK4.