dialog-notebook: Fix fix for saving child padding…
…namely that we need to save the ChildPropertyProxy.get_value(), NOT the
proxy itself as `auto` here did - so we save the value the child prop
HAD, not whatever gibberish we'll get from the proxy after it's removed.
This manifested in one of my other WIP branches as odd wrong prop values

So, rather than having to explicitly convert - which might be overlooked
later as I did here - use `auto` BUT explicitly on current .get_value().