Fix style control in object to path and stroke to path
tspans inside of tspans did not have their style respected when creating
the new style for the new path. tspan parent styles must be merged.

The stroke to path would produce a copy of the fill path (not the stroke)
if the stroke color was set, but the stroke width was zero. This is not
correct and instead if there's no path it should return.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4403