ui/util + for_each_child taking functor+cont/break
Add this helper, which may seem trivial now, but will be of great use
once migrating to GTK4, which drops GtkContainer and moves children to
GtkWidget, AND changes how to navigate along the list of their children.
Thus if we can to calling this, those sites will then become GTK4-proof.