ColumnMenuBuilder: from Gtk::Menu* to Popover, etc
and in replacing Menu/MenuItem with other widgets, we basically need to
change those widgets' CSS names to menu and menuitem, so theming won't
get broken (or become really laborious to fix ourselves, & even then not
work for users of other themes) - so this uses custom widgets and
glibmmʼs ExtraClassInit to make it so. Add a glibmm CMake dep to match.

Worth noting here is that, to look like GtkMenu/Item, I must do a bit of
hackery to make focused items appear hovered. In style.css the outline
is suppressed accordingly. We also un-hover when the key focus changes,
so we don't end up with 2 items highlighted in blue. Lastly style.css
gets a bit to squish unwanted popover theming and look more like a Menu.

This was initially done in DialogNotebook only, but then mkov pointed
out that ColumnMenuBuilder was nearly the same thing. So this makes
Notebook use ColumnMenuBuilder, giving the latter a couple of new,
optional constructor arguments to make it work and keep the same look.

The CompletionPopup then has its Menu removed from the UI file as it
needs to change to our PopoverMenu and that's unnecessary hassle to make
usable with Gtk::Builder, when we simply don't have to. The other user
of that is FilterEffectsDialog which, amazingly, is happy with NO edit!!

Doing raw new & delete on Notebookʼs PopoverMenuItems caused a crash via
reload_tab_menu(), in PopoverMenu::attach()'s dynamic_cast, which makes
NO sense to me. Storing them as unique_ptrs solved that and seems nicer
(if a bit more manual bookkeeping) and more explicit about what we're
doing. Note also that we must remove() in gtkmm4, as destroying a C++
widget in a parent no longer nukes the C widget… so start doing so now.

Lastly, this adds utils.h: for_each_parent(), and generalises the enum
class ForEachChildResult to be called ForEachResult so both can use it.

https://gitlab.com/inkscape/inkscape/-/merge_requests/5512