popover-menu: Bin workaround that's doing nothing…
…another leftover. It seems whatever it had been working around was just
a quirk that I had introduced, but which is no longer relevant, since I
see the exact same behaviour with or without this code: the notebook
action menu has the correct 1st item focused on 1st popup, whereas the
notebook tab menus have nothing, and the filter-effects-dialog has the
bottom-left. I cannot seem to conjure any workaround that affects this.
So, for now, just remove the misleading/no-op code and leave a FIXME.