preGTK4: color-item: GtkMenu to Gio::Menu/Popover
and during this, cleanup the code creating the menu, sort gradient names
for which we offer Convert, and use Title Case where missing from items.

This also hoists the hackery that made items in our PopoverMenu act like
those in GtkMenu, to its own file menuize.[h|cpp] and uses that here, as
well as adding some CSS to make GtkPopoverMenu items look like GtkMenuʼs
– more so than previously, by making corners of Popover square like Menu
as that gets it closer again to what users are familiar with from <=1.3.

N.B.: Gtk::ModelButton appears to reveal that the hackery w/ prelight is
required on ::enter too, not just ::motion, else we get small slivers of
space at the edge of items where previously hovered will still be prelit

https://gitlab.com/inkscape/inkscape/-/merge_requests/5592