Fix bleed and margin unit issues in document
The margin and bleed were being stored in the svg with units. For example mm or inches
This is incorrect since all units stored in the svg must be user units.

In order to preserve backwards compatibility with documents created with Inkscape 1.3
we pass the document scale into our reading methods so they will be converted safely
to user units on load.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/4487