SBToolItem: mv file static Popover → self-destruct
As Rafał pointed out during review, this is not exactly easy to
understand at a first glance. I had made it static in part because I had
found that making it self-destruct on ::closed resulted in errors, but
that appears no longer to be the case as Popover::closed works fine now.
However glancing at GTK3 code, I think Widget::hide is a bit safer as it
seems to run a bit late, so less likely to interfere with GTK stuff that
runs on close or hide.

The self-destruct behaviour is implemented in ui/popup-menu.* as the
function on_hide_reset(), which takes a shared_ptr to a Widget, & copies
a shared_ptr to said widget into a ::hide handler that will reset() it &
hence cause the destruction of the widget once it has been hidden/closed