tool-base—Donʼt warn of unhandled rotate if locked
In that case we know how to handle the event; we just choose not to
honour it, since the user has requested to lock the canvas for rotation.
So we should probably not print "unhandled event" and indicate there's a
problem, when there isn't: we are happy with the event and to ignore it.