ui/util: Fix still recursed if were asked not to!
Add an int level argument to track whether we are on the root level, as
we always need to 'recurse' from that, but may not want to if we are >1

Rename `test_self` to `plus_self`, seems a somewhat better name to me.

This was clearly just wrong and had visible results, e.g. the XML/Style
dialog got empty space at the bottom because we failed to move children
from the dialog to the wrapperbox in the (eek! scary!) dialog-notebook.