Ctrlr: + FIXME. PrefWidget: − unused/depr includes
The Controller change relates to my previous fix for signals that take
no result but for which the concept was letting us provide functions
that DO return non-void. We should fix it to diagnose that at the
source. I just didn't yet manage to find anything that didn't error out.

PreferencesWidget was including Gtk::Frame but not using, and Gtk::Table
which is deprecated in GTK3 already, and used nowhere in Inkscape now.