ColorWheels: Stop using Style properties & .render
StyleContext.render_focus() is gone in GTK4, as are style properties, so
thankfully we don't really need them here & can even do better actually!

* Make the constants constexpr.
* Introduce new constants for the marker radius and focus margin/dashing
  ...some of which replace use of GTK style properties ie a giant no-no!
* If the ring has focus, instead of making GTK render focus around whole
  rectangular allocation, draw our own focus ring round our outer circle
* Make said focus ring dashed, consistent with (most) GTK focus outlines
* Make the existing focus rings around markers dashed, exactly the same.
* Use marker_radius where it's = but semantically better than other vars
* Use get_width|height() where we don't need GTK4-deprd get_allocation()
* Add comment re a similar-but-different function maybe worth pondering.

***

ColorWheel: Be abstract, not providing dummy impls

Instead of having virtual functions that do nothing in the base, make
those pure virtual, so they must be implemented by derived classes.

Except _set_from_xy(), which the base does not use, so that should be
moved to non-inherited functions in each of the subclasses that use it.