ColorPalette: Fix popup showed wrong selected item
The CustomMenuItem was meant to store the id, as member `tag, but it never actually bothered to set that member, so our search for an item with matching tag to select failed, and we left the last one selected. Fix that, and rename the member to `id\` to match the calling location.

The loop removing children was overcomplicated, so just store a vector of our CustomMenuItems, renamed to ColorPaletteMenuItem to avoid clash, and remove all of those, which will leave the 2 constant ones at bottom.

Explain we prepend in reverse to go above said items & be const-correct.

Use get_width|height() instead of get_allocation(), as !deprecated@GTK4.

plus usual cleanups