ui/manage—Fix assumption that refs/ptrs stay valid
I had wrongly assumed that the guarantee of std::[multi]map that
references/pointers/iterators to elements are never invalidated (well,
only by extraction to node handles) was also applicable to
std::unordered_[multi]map, but that is not the case because the latter
may rehash and that invalidates previous addresses of elements. So, this
needs to use an ordered multimap so that popup-menu.cpp can always
access the managed slot at the same address, instead of invoking UB if
the unordered_multimap got big enough to rehash at some future point.